+++ HISTORY.md
@@ -1,8 +1,31 @@
-# v0.41
+# 0.41
 
 Removed the `varinfo` keyword argument from `DynamicPPL.TestUtils.AD.run_ad` and replaced the `varinfo` field in `ADResult` with `ldf::LogDensityFunction`.
 
 Removed `getargnames`, `getmissings`, and `Base.nameof(::Model)` from the public API (export and documentation) as they are considered internal implementation details.
+
+# 0.40.17
+
+Implemented missing methods for `Base.copy` on internal structs.
+
+# 0.40.16
+
+Fixed `Base.copy` for `VNTAccumulator` and `TSVNTAccumulator` to also copy the `acc.f` field, not just `acc.values`.
+This makes sure that the accumulator is thread-safe if `acc.f` contains some mutable state.
+
+# 0.40.15
+
+DynamicPPL now allows you to set the type that log-probabilities are initialised with, using the `DynamicPPL.set_logprob_type!` function.
+This records a compile-time preference so requires restarting Julia to take effect.
+
+This allows model log-probability accumulation to work with different numerical precisions.
+For example, if your model is defined using distributions that are parameterised by `Float32` only (and avoid promoting them to `Float64` elsewhere in the model), and you call `DynamicPPL.set_logprob_type!(Float32)`, the resulting log-probabilities will also be `Float32`.
+
+Previously, DynamicPPL would automatically choose a `Float64` log-probability, causing any lower-precision model to be promoted.
+
+The function `DynamicPPL.get_input_vector_type(::LogDensityFunction)` is now exported, in order to help with querying the type that log-probabilities are initialised with.
+
+`DynamicPPL.typed_identity` is deprecated; please use `Bijectors.VectorBijectors.TypedIdentity()` instead (it does the same thing).
 
 # 0.40.14
 
@@ -32,8 +55,6 @@ Added more docs on special VNT operations, namely `densify!!` and `skeleton`.
 # 0.40.8
 
 Added the `skeleton(::VarNamedTuple)` function, which creates a completely blank VarNamedTuple except for any PartialArrays, which are retained with their original shape and type but with all elements set to `nothing`.
-
-> > > > > > > main
 
 # 0.40.7